requirements.txt
dill==0.3.5.1
from-root==1.0.2
gdown
flask-cors
flask
scikit-learn
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics>=8.0.111
joblib

pandas>=1.1.4
seaborn>=0.11.0

setuptools>=65.5.1


# -e .